% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{get_species}
\alias{get_species}
\title{This function returns the species column from the Palmer penguins dataset}
\usage{
get_species()
}
\value{
dataframe of 1 column from the palmer penguins dataframe
}
\description{
This function returns the species column from the Palmer penguins dataset
}
